style(transit-map): modify the function structure
- 함수 구조 수정
